+++ lib/src/clojure_lsp/feature/document_symbol.clj
@@ -1,7 +1,9 @@
 (ns clojure-lsp.feature.document-symbol
   (:require
+   [clojure-lsp.parser :as parser]
    [clojure-lsp.queries :as q]
-   [clojure-lsp.shared :as shared]))
+   [clojure-lsp.shared :as shared]
+   [rewrite-clj.zip :as z]))
 
 (set! *warn-on-reflection* true)
 
@@ -29,6 +31,7 @@
     :var-usages (if (:defmethod el)
                   :function
                   :variable)
+    :keyword-usages :field
     :null))
 
 (defn element->name [{elem-name :name :keys [dispatch-val-str]}]
@@ -49,19 +52,53 @@
 (defn ^:private symbol-order [{:keys [selection-range]}]
   [(:line (:start selection-range)) (:character (:start selection-range))])
 
+(defn ^:private remove-first [item-to-remove coll]
+  (if (empty? coll)
+    ()  ; literal name for empty list
+    (let [item (first coll)]
+      (if (= item item-to-remove)
+        (rest coll)
+        (cons (first coll) (lazy-seq (remove-first item-to-remove (rest coll))))))))
+
+(defn ^:private edn->element-tree [m keyword-elements]
+  (sort-by symbol-order
+           (reduce
+             (fn [acc [k v]]
+               (let [element (first (filter #(= (:name %) (name k)) keyword-elements))
+                     document-symbol (element->document-symbol element)
+                     kind (cond
+                            (string? v) :string
+                            (keyword? v) :field
+                            (boolean? v) :boolean
+                            (vector? v) :array
+                            (set? v) :array
+                            :else :struct)
+                     document-symbol (assoc document-symbol :kind kind)]
+                 (if (map? v)
+                   (conj acc (assoc document-symbol
+                                    :children
+                                    (edn->element-tree v (remove-first element keyword-elements))))
+                   (conj acc document-symbol))))
+             []
+             m)))
+
 (defn document-symbols [db uri]
   (let [namespace-definition (q/find-namespace-definition-by-uri db uri)]
-    [{:name (or (some-> namespace-definition :name name)
-                ;; TODO Consider using URI for display purposes, especially if
-                ;; we support remote LSP connections
-                (shared/uri->filename uri))
-      :kind (element->symbol-kind namespace-definition)
-      :range shared/full-file-range
-      :selection-range (if namespace-definition
-                         (shared/->scope-range namespace-definition)
-                         shared/full-file-range)
-      :children (->> (concat (q/find-var-definitions db uri true)
-                             (q/find-defmethods db uri))
-                     (map element->document-symbol)
-                     (sort-by symbol-order)
-                     vec)}]))
+    (if (identical? :edn (shared/uri->file-type uri))
+      (-> (parser/zloc-of-file db uri)
+          z/sexpr
+          (edn->element-tree (get-in db [:analysis uri :keyword-usages])))
+      [{:name (or (some-> namespace-definition :name name)
+                  ;; TODO Consider using URI for display purposes, especially if
+                  ;; we support remote LSP connections
+                  (shared/uri->filename uri))
+        :kind (element->symbol-kind namespace-definition)
+        :range shared/full-file-range
+        :selection-range (if namespace-definition
+                           (shared/->scope-range namespace-definition)
+                           shared/full-file-range)
+        :children (->> (concat (q/find-var-definitions db uri true)
+                               (q/find-defmethods db uri))
+                       (map element->document-symbol)
+                       (sort-by symbol-order)
+                       vec)}])))