+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - Add new setting to change diagnostics range type. #1530
   - Spec compliance: avoid registering client capabilities if dynamic registration is not set. #1535
   - Improve java support: using new `java-member-definitions` analysis, first feature is the allow completion of static members (fields and methods) for java classes.
+  - Show edn tree when in edn files via `documentSymbol` feature.
 
 ## 2023.02.27-13.12.12